Split database code into its own module
