+++ src/constants.py
@@ -6,3 +6,5 @@
 FLASK_CONTAINER_NAME = "flask-app"
 FLASK_SERVICE_NAME = "flask-app"
 FLASK_ENV_CONFIG_PREFIX = "FLASK_"
+FLASK_DATABASE_NAME = "flask-app"
+FLASK_SUPPORTED_DATABASES = ("mysql", "postgresql")
